Using Monte Carlo method, compute the area.py
compute the area of the following set:
􏰂(x+3)2 +y2 +􏰂x2 +y2 +􏰂(x−3)2 +y2 ≤17
with the accuracy of at least 1%. Explain why you believe that you achieved this accuracy. Check yourself! Repeat the entire analysis 100 times, and plot the histogram of the results. Are they within 1% of each other? Include the plot of the set defined above with your submission.