Create Tic-tac-toe.2
a program that plays a game of m, n, k-Tic-tac-toe with a user. At the beginning the user chooses the values of m, n, k, and the type of AI that they will play against. Each round, the game prints our the state of the board, asks the user where they would like to place their mark, and implements this decision. Once one of the player won, the program declares the result and asks if the user would like to continue. The first player is selected at random. Here, m is the number of rows, n is the number of columns, and k is the length of the sequence of marks arranged horizontally, vertically, or diagonally that constitutes a win.
For this problem we will implement three types of AI:
• AI Level 0. The program then places its own mark on a randomly chosen available position.
• AI Level 1. The program checks if it can win by placing its mark in any of the available positions. If so, it makes the winning move. Otherwise, it places its mark on a randomly chosen available position.
• AI Level 2. The program checks if it can win by placing its mark in any of the available positions. If so, it makes the winning move. If not, it checks if the opponent can win the game on the next move by placing their mark on one of the available positions. If so, it blocks that position with its own mark. Otherwise, it places its mark on a randomly chosen available position.